+++ src/packages/table/demos/taro/demo6.tsx
@@ -1,0 +1,46 @@
+import React, { useState } from 'react'
+import { Table } from '@nutui/nutui-react-taro'
+
+interface TableColumnProps {
+  key: string
+  title?: string
+  align?: string
+  sorter?: ((a: any, b: any) => number) | boolean | string
+  render?: (rowData: any, rowIndex: number) => string | React.ReactNode
+  fixed?: 'left' | 'right'
+  width?: number
+}
+
+const Demo6 = () => {
+  const [columns1, setColumns1] = useState<Array<TableColumnProps>>([
+    {
+      title: '姓名',
+      key: 'name',
+    },
+    {
+      title: '性别',
+      key: 'sex',
+      render: (record: any) => {
+        return (
+          <span style={{ color: record.sex === '女' ? 'blue' : 'green' }}>
+            {record.sex}
+          </span>
+        )
+      },
+    },
+    {
+      title: '学历',
+      key: 'record',
+    },
+  ])
+
+  const [data2, setData2] = useState([])
+
+  return (
+    <>
+      <Table columns={columns1} data={data2} />
+      <Table columns={columns1} data={data2} noData="这里是自定义展示" />
+    </>
+  )
+}
+export default Demo6
