+++ src/packages/table/demos/taro/demo7.tsx
@@ -1,0 +1,71 @@
+import React, { useState } from 'react'
+import { Table, Button } from '@nutui/nutui-react-taro'
+import { Star } from '@nutui/icons-react-taro'
+
+const Demo7 = () => {
+  const [columns4, setColumns4] = useState([
+    {
+      title: '姓名',
+      key: 'name',
+      align: 'center',
+    },
+    {
+      title: '性别',
+      key: 'sex',
+    },
+    {
+      title: '学历',
+      key: 'record',
+    },
+    {
+      title: '操作',
+      key: 'render',
+    },
+  ])
+
+  const [data4, setData4] = useState([
+    {
+      name: 'Tom',
+      sex: '男',
+      record: '小学',
+      render: () => {
+        return (
+          <Button
+            onClick={() => Toast.show('hello')}
+            size="small"
+            type="primary"
+          >
+            <div>Hello</div>
+          </Button>
+        )
+      },
+    },
+    {
+      name: 'Lucy',
+      sex: '女',
+      record: '本科',
+      render: () => {
+        return <Star size="14px" />
+      },
+    },
+    {
+      name: 'Jack',
+      sex: '男',
+      record: '高中',
+      render: () => {
+        return (
+          <Button
+            type="success"
+            size="small"
+            onClick={() => window.open('https://www.jd.com')}
+          >
+            <div>跳转到京东</div>
+          </Button>
+        )
+      },
+    },
+  ])
+
+  return <Table columns={columns4} data={data4} />
+}
+export default Demo7
